LayoutTests/http/tests/from-origin/redirect-script-from-origin-same-blocked-expected.txt

CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
CONSOLE MESSAGE: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
Tests that a redirected cross-origin script load fails if the server blocks it with a 'From-Origin: same' response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Script load blocked.
PASS successfullyParsed is true

TEST COMPLETE
